helperai/requirements.txt
beautifulsoup4==4.12.3
faiss-cpu==1.7.4
matplotlib==3.8.4
numpy==1.26.4
ollama==0.2.0
pandas==2.2.2
pdfplumber==0.11.4
python-docx==1.1.2
requests==2.32.3
sentence-transformers==3.0.1
streamlit==1.36.0
